+++ internal/cards/dictionary.go
@@ -5,6 +5,8 @@ import (
 	"os"
 
 	"foosoft.net/projects/jmdict"
+	"github.com/ikawaha/kagome-dict/ipa"
+	"github.com/ikawaha/kagome/v2/tokenizer"
 	"github.com/mochi-co/kana-tools"
 )
 
@@ -41,6 +43,7 @@ func (cd *CardData) LoadDictionary() {
 	var kanjiMap = make(map[string][]*jmdict.JmdictEntry)
 	var readingMap = make(map[string][]*jmdict.JmdictEntry)
 	var nonKanjiReadingMap = make(map[string][]*jmdict.JmdictEntry)
+	var meaningMap = make(map[string][]*jmdict.JmdictEntry)
 	for i := range dict.Entries {
 		entry := &dict.Entries[i]
 		for _, kanji := range entry.Kanji {
@@ -64,6 +67,12 @@ func (cd *CardData) LoadDictionary() {
 			}
 			readingMap[kana.ToHiragana(reading.Reading)] = append(readingMap[reading.Reading], entry)
 		}
+
+		for _, sense := range entry.Sense {
+			for _, gloss := range sense.Glossary {
+				meaningMap[gloss.Content] = append(meaningMap[gloss.Content], entry)
+			}
+		}
 	}
 
 	log.Printf("Index built")
@@ -71,6 +80,7 @@ func (cd *CardData) LoadDictionary() {
 	cd.DictionaryKanjiMap = kanjiMap
 	cd.DictionaryReadingMap = readingMap
 	cd.DictionaryNonKanjiReadingMap = nonKanjiReadingMap
+	cd.DictionaryMeaningMap = meaningMap
 	cd.DictionaryEntities = entities
 }
 
@@ -116,6 +126,60 @@ func (t *Token) AddDictionaryEntry(cd *CardData) {
 	}
 }
 
+func SearchDictionary(cd *CardData, query string) DictionarySearchData {
+	// If query is in romanji, convert it to hiragana
+	originalQuery := query
+	if !kana.ContainsHiragana(query) && !kana.ContainsKatakana(query) && !kana.ContainsKanji(query) {
+		query = kana.ToHiragana(query)
+	}
+
+	t, err := tokenizer.New(ipa.Dict(), tokenizer.OmitBosEos())
+	if err != nil {
+		panic(err)
+	}
+
+	tokens := t.Analyze(query, tokenizer.Normal)
+	log.Printf("Found %d tokens", len(tokens))
+
+	var result []DictionaryEntry
+
+	// Search on the whole query before searching on the tokenized parts
+	result = append(result, GetDictionaryEntries(query, cd.DictionaryKanjiMap)...)
+	result = append(result, GetDictionaryEntries(query, cd.DictionaryNonKanjiReadingMap)...)
+	result = append(result, GetDictionaryEntries(query, cd.DictionaryReadingMap)...)
+
+	// Search on the original query to search English meanings
+	result = append(result, GetDictionaryEntries(originalQuery, cd.DictionaryMeaningMap)...)
+
+	var tokenStrings []string
+	for _, token := range tokens {
+		to := ConvertToken(token)
+		result = append(result, GetDictionaryEntries(to.BaseForm, cd.DictionaryKanjiMap)...)
+		result = append(result, GetDictionaryEntries(to.BaseForm, cd.DictionaryNonKanjiReadingMap)...)
+		result = append(result, GetDictionaryEntries(to.Pronunciation, cd.DictionaryReadingMap)...)
+		if to.BaseForm != "" {
+			tokenStrings = append(tokenStrings, to.BaseForm)
+		}
+	}
+
+	return DictionarySearchData{
+		DictSearchTerm:    originalQuery,
+		DictSearchResults: result,
+		Tokens:            tokenStrings,
+	}
+}
+
+func GetDictionaryEntries(term string, dict map[string][]*jmdict.JmdictEntry) []DictionaryEntry {
+	var result []DictionaryEntry
+
+	entries := dict[term]
+	for _, entry := range entries {
+		result = append(result, convertJmdictEntryToDictionaryEntry(*entry))
+	}
+
+	return result
+}
+
 // Convert a JMdict entry to a dictionary entry for ease of use
 func convertJmdictEntryToDictionaryEntry(entry jmdict.JmdictEntry) DictionaryEntry {
 	var de DictionaryEntry
@@ -134,6 +198,7 @@ func convertJmdictEntryToDictionaryEntry(entry jmdict.JmdictEntry) DictionaryEnt
 		var dd DictionaryDefinition
 
 		dd.PartsOfSpeech = sense.PartsOfSpeech
+		dd.PartsOfSpeech = append(dd.PartsOfSpeech, sense.Misc...)
 		for _, gloss := range sense.Glossary {
 			dd.Definitions = append(dd.Definitions, gloss.Content)
 		}