+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.2.1 - 2023-04-16
+### Dictionary Search Page
+`/dictionarysearch` is now a thing. It's a page that allows you to search for words in the dictionary.
+
 ## 0.2.0 - 2023-04-15
 ### Dictionary Popup
 The text analysis page now has a dictionary popup for words that do not exist as cards.